requirements.txt
anthropic>=0.18.0
openai>=1.12.0
pyyaml>=6.0
